common/engines/yolov4/requirements.txt
numpy
torch
tensorboardX
scikit-image
matplotlib
tqdm
easydict
Pillow
opencv-python
pycocotools
